Don't append single element at the end of the list
In erlang, lists are linked lists, complexity of placing element at the end of the list is O(N) where N is length of the list.
It's more efficient to put element in head (O(1)), and later reverse whole list.

Maybe here it won't be noticeable, but when working with big lists, it will a big payoff.